refactor and moving
