+++ src/index.js
@@ -7,65 +7,16 @@ if (!window.__loaded) {
   document.write('<script src="../lib/blocks.js"></script>')
   document.write('<script src="../lib/javascript-commands.js" onload="startApp()"}></script>')
 }
-
 window.__loaded = true
 
 const _ = require('lodash')
 const choo = require('choo')
-const update = require('immutability-helper')
-const mainView = require('./main')
-const interpreter = require('./interpreter')
-
 const app = choo()
 
-app.model({
-  state: {
-    robot: { x: 5, y: 5 },
-    running: false,
-    srcCode: ''
-  },
-
-  effects: {
-    runCode: (data, state, send, done) => {
-      interpreter.run(state.srcCode, send, done)
-      console.log(state.srcCode)
-    }
-  },
-
-  reducers: {
-    move: ({ direction }, state) => {
-      let { robot: { x, y } } = state
-
-      switch (direction) {
-        case 'UP':
-          y = y - 10
-          break
-        case 'DOWN':
-          y = y + 10
-          break
-        case 'LEFT':
-          x = x - 10
-          break
-        case 'RIGHT':
-          x = x + 10
-          break
-      }
-
-      return update(state, {
-        robot: {
-          x: { $set: x },
-          y: { $set: y }
-        }
-      })
-    },
-
-    changeRunningState: ({ running }, state) => update(state, { running: { $set: running } }),
-    updateCode: ({ srcCode }, state) => update(state, { srcCode: { $set: srcCode } })
-  }
-})
+app.model(require('./model'))
 
 app.router((route) => [
-  route('/', mainView)
+  route('/', require('./components/main'))
 ])
 
 // setup hot module replacement
